requirements.txt
# Audio Processing
sounddevice>=0.4.6
pyaudio>=0.2.13
pulsectl>=23.5.2
librosa>=0.10.1
pydub>=0.25.1
webrtcvad>=2.0.10
scipy>=1.11.0
numpy>=1.24.0

# Speech-to-Text
openai-whisper>=20231117
torch>=2.0.0
torchaudio>=2.0.0

# Speaker Diarization
pyannote.audio>=3.1.0
speechbrain>=0.5.16

# NLP and Analysis
spacy>=3.7.0
nltk>=3.8.1
transformers>=4.35.0
scikit-learn>=1.3.0
textstat>=0.7.3

# Database
sqlalchemy>=2.0.0
alembic>=1.12.0

# CLI Interface
click>=8.1.7
rich>=13.7.0
tabulate>=0.9.0

# Web UI (Future)
flask>=3.0.0
flask-sqlalchemy>=3.1.0
flask-cors>=4.0.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0.1
python-dateutil>=2.8.2
